http-adapters: move shared JSON parsing and payload locating into http-adapter.c
Both splunk-adapter.c and openobserve-adapter.c contained identical copies
of _parse_response_json() and _locate_offending_payload().  Consolidate
them into http_adapter_parse_response_json() and
http_adapter_locate_offending_payload() in the base module so adapters
don't each carry their own copy.
